requirements.txt
pydantic-settings
python-dotenv
sqlalchemy
psycopg2-binary
alembic
uvicorn
fastapi
email-validator
passlib[bcrypt]
bcrypt==4.0.1